Move main component content
Prepare for adding other pages by moving map and sidebar to MainPage component.
